requirements.txt
python-dotenv
fastapi 
uvicorn 
PyMuPDF 
SQLAlchemy 
psycopg2-binary
pymysql
websockets 
langchain
langchain-community
python-multipart
pytest 
httpx
pytest-asyncio
chromadb
pytest-timeout
groq
langchain-groq
langchain_google_genai